Factor http functionality into a package
