+++ testconfig.yaml
@@ -5,13 +5,15 @@ jobs:
         - type: http
           args:
             count: 1
-            method: GET
-            path: 'https://localhost:8080/search?searchid={{ index (.Value (ctx_key "config")) "args" "jobs" 0 "type" }}&l10n=ru&reqenc=&text={{ random_uuid }}'
+            request:
+              method: GET
+              path: 'https://localhost:8080/search?searchid={{ index (.Value (ctx_key "config")) "args" "jobs" 0 "type" }}&l10n=ru&reqenc=&text={{ random_uuid }}'
             client:
               timeout: 1s
         - type: http
           args:
-            method: GET
-            path: "https://localhost:8090/search?searchid={{ random_uuid }}&l10n=ru&reqenc=&text={{ random_uuid }}"
+            request:
+              method: GET
+              path: "https://localhost:8090/search?searchid={{ random_uuid }}&l10n=ru&reqenc=&text={{ random_uuid }}"
             client:
               proxy_urls: '{{ join get_proxylist "," }}'
